requirements.txt
celery
asgiref
Django
django-cors-headers
djangorestframework
pytz
sqlparse
